Cleaned up code and moved the ChatClient to the tests directory.
