+++ backend/src/main/java/com/angelomelonas/grpcwebchat/grpc/ChatImpl.java
@@ -2,7 +2,9 @@ package com.angelomelonas.grpcwebchat.grpc;
 
 import com.angelomelonas.grpcwebchat.ChatGrpc;
 import com.angelomelonas.grpcwebchat.ChatOuterClass.Message;
-import com.angelomelonas.grpcwebchat.ChatOuterClass.SendMessageResponse;
+import com.angelomelonas.grpcwebchat.ChatOuterClass.MessageRequest;
+import com.angelomelonas.grpcwebchat.ChatOuterClass.MessagesRequest;
+import io.grpc.stub.StreamObserver;
 import org.lognet.springboot.grpc.GRpcService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -10,43 +12,41 @@ import org.slf4j.LoggerFactory;
 import java.time.Instant;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.UUID;
 
 @GRpcService
 public class ChatImpl extends ChatGrpc.ChatImplBase {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ChatImpl.class);
-    private Map<String, Message> messages;
+    private Map<Long, Message> messages = new HashMap<>();
 
-    public ChatImpl() {
-        messages = new HashMap<String, Message>();
-    }
+    private Long idCounter = 0L;
 
     @Override
-    public void sendMessage(com.angelomelonas.grpcwebchat.ChatOuterClass.MessageRequest request, io.grpc.stub.StreamObserver<com.angelomelonas.grpcwebchat.ChatOuterClass.SendMessageResponse> responseObserver) {
-        LOGGER.info("Received client message {}", request);
+    public void sendMessage(MessageRequest request, StreamObserver<Message> responseObserver) {
+        LOGGER.info("Received client message: {} from user: {}", request.getMessage(), request.getUsername());
 
+        // Get the user and message details sent from the client.
         String receivedMessage = request.getMessage();
         String username = request.getUsername();
         Long timestamp = Instant.now().toEpochMilli();
 
+        // Create a new message.
         Message message = Message.newBuilder().setMessage(receivedMessage).setUsername(username).setTimestamp(timestamp).build();
 
-        String messageId = UUID.randomUUID().toString();
+        // Store the message on the server side.
+        messages.put(idCounter++, message);
 
-        messages.put(messageId, message);
-
-        // TODO: Reply that the message was sent successfully.
-        responseObserver.onNext(SendMessageResponse.newBuilder().build());
+        // Respond with the newly created message.
+        responseObserver.onNext(message);
         responseObserver.onCompleted();
     }
 
     @Override
-    public void getMessages(com.angelomelonas.grpcwebchat.ChatOuterClass.GetMessagesRequest request, io.grpc.stub.StreamObserver<com.angelomelonas.grpcwebchat.ChatOuterClass.Message> responseObserver) {
+    public void getMessages(MessagesRequest request, StreamObserver<Message> responseObserver) {
         LOGGER.info("Client requested messages {}", request);
 
+        // Stream all messages to the client.
         messages.forEach((id, message) -> responseObserver.onNext(message));
         responseObserver.onCompleted();
     }
-
 }
